PEPPER-1613 LMS adult consent version 3
